requirements-310-backup.txt
alembic==1.16.5
annotated-types==0.7.0
ansicolors==1.1.8
anyio==4.10.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
Authlib==1.6.5
autopep8==2.3.2
babel==2.17.0
backports-datetime-fromisoformat==2.0.3
bandit==1.8.6
beautifulsoup4==4.13.5
black==25.9.0
bleach==6.2.0
blis==0.7.11
catalogue==2.0.10
cattrs==25.2.0
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
cloudscraper==1.2.71
comm==0.2.3
confection==0.1.5
coverage==7.10.6
cramjam==2.11.0
cryptography==46.0.2
cymem==2.0.11
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
dparse==0.6.4
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.117.1
fastjsonschema==2.21.2
fastparquet==2024.11.0
feedparser==6.0.12
filelock==3.19.1
flake8==7.3.0
fqdn==1.5.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
identify==2.6.14
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython==8.37.0
isoduration==20.11.0
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab>=4.4.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
lxml==6.0.1
lxml_html_clean==0.4.2
Mako==1.3.10
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.4
mpmath==1.3.0
murmurhash==1.0.13
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
newspaper4k==0.9.3.1
nltk==3.9.1
nodeenv==1.9.1
notebook_shim==0.2.4
numpy==1.26.4
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
papermill==2.6.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
preshed==3.0.10
prettytable==3.16.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.4
PySocks==1.7.1
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-cache==1.2.1
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
ruff==0.13.2
safetensors==0.6.2
safety==3.6.2
safety-schemas==0.0.16
scikit-learn==1.5.1
scipy==1.15.3
selenium==4.35.0
selenium-stealth==1.0.6
Send2Trash==1.8.3
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
skops==0.13.0
smart_open==7.3.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
stack-data==0.6.3
starlette==0.48.0
stevedore==5.5.0
storysniffer==1.0.7
structlog==25.4.0
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.6.0
tinycss2==1.4.0
tldextract==5.3.0
tokenizers==0.22.1
tomli==2.2.1
tomlkit==0.13.3
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
trio==0.30.0
trio-websocket==0.12.2
typer==0.17.4
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
undetected-chromedriver==3.5.5
uri-template==1.3.0
url-normalize==2.2.1
urllib3==2.5.0
uvicorn==0.36.0
virtualenv==20.34.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
